Added solution to median.py

+++ median.py
@@ -11,7 +11,7 @@ while True:
     else:
         break
 if(len(numbers) % 2 == 0):
-    print((numbers[len(numbers/2)] + len(numbers/2 + 1))/2)
+    print((int(numbers[len(numbers/2)]) + int(len(numbers/2 + 1)))/2)
 else:
     print(numbers[int(len(numbers)/2)])
